Add a StreamVisitor which is able to handle a newly created Stream prior to any StreamHandler
