+++ src/main/kotlin/io/libp2p/mux/MuxHandler.kt
@@ -3,6 +3,7 @@ package io.libp2p.mux
 import io.libp2p.core.Stream
 import io.libp2p.core.StreamHandler
 import io.libp2p.core.StreamPromise
+import io.libp2p.core.StreamVisitor
 import io.libp2p.core.mux.StreamMuxer
 import io.libp2p.etc.CONNECTION
 import io.libp2p.etc.STREAM
@@ -20,12 +21,13 @@ import java.util.concurrent.atomic.AtomicLong
 open class MuxHandler(
     private val ready: CompletableFuture<StreamMuxer.Session>?,
     inboundStreamHandler: StreamHandler<*>,
+    private val streamVisitor: StreamVisitor?
 ) : AbstractMuxHandler<ByteBuf>(), StreamMuxer.Session {
     private val idGenerator = AtomicLong(0xF)
 
 
     override val inboundInitializer: MuxChannelInitializer<ByteBuf> = {
-        inboundStreamHandler.handleStream(createStream(it))
+        createAndHandleStream(it, inboundStreamHandler)
     }
 
     override fun handlerAdded(ctx: ChannelHandlerContext) {
@@ -79,10 +81,20 @@ open class MuxHandler(
         return stream
     }
 
+    private fun <T> createAndHandleStream(
+        channel: MuxChannel<ByteBuf>,
+        protocolStreamHandler: StreamHandler<T>
+    ): CompletableFuture<out T> {
+        val stream = createStream(channel)
+        streamVisitor?.onNewStream(stream)
+        return protocolStreamHandler.handleStream(stream)
+    }
+
     override fun <T> createStream(streamHandler: StreamHandler<T>): StreamPromise<T> {
         val controller = CompletableFuture<T>()
-        val stream = newStream { streamHandler.handleStream(createStream(it)).forward(controller) }
-            .thenApply { it.attr(STREAM).get() }
+        val stream = newStream {
+            createAndHandleStream(it, streamHandler).forward(controller)
+        }.thenApply { it.attr(STREAM).get() }
         return StreamPromise(stream, controller)
     }
 }
