+++ src/test/kotlin/io/libp2p/mux/MultiplexHandlerTest.kt
@@ -4,6 +4,7 @@ import io.libp2p.core.ConnectionClosedException
 import io.libp2p.core.Libp2pException
 import io.libp2p.core.Stream
 import io.libp2p.core.StreamHandler
+import io.libp2p.core.StreamVisitor
 import io.libp2p.etc.types.fromHex
 import io.libp2p.etc.types.getX
 import io.libp2p.etc.types.toByteArray
@@ -50,7 +51,7 @@ class MultiplexHandlerTest {
                 childHandlers += handler
             }
         )
-        multistreamHandler = object : MuxHandler(null, streamHandler) {
+        multistreamHandler = object : MuxHandler(null, streamHandler, null) {
             // MuxHandler consumes the exception. Override this behaviour for testing
             override fun exceptionCaught(ctx: ChannelHandlerContext, cause: Throwable) {
                 super.exceptionCaught(ctx, cause)
@@ -190,7 +191,8 @@ class MultiplexHandlerTest {
         ech.close().await()
 
         val staleStream =
-            multistreamHandler.createStream(StreamHandler.create { println("This shouldn't be displayed: parent stream is closed") })
+            multistreamHandler.createStream(StreamVisitor
+            { println("This shouldn't be displayed: parent stream is closed") }.toStreamHandler())
 
         assertThrows(ConnectionClosedException::class.java) { staleStream.stream.getX(3.0) }
     }
